+++ src/index.js
@@ -1,4 +1,4 @@
-const { scan, scanBuffer } = require('./ClamAVScanner.js');
-const { Verdict }          = require('./verdicts.js');
+const { scan, scanBuffer, scanStream } = require('./ClamAVScanner.js');
+const { Verdict }                       = require('./verdicts.js');
 
-module.exports = { scan, scanBuffer, Verdict };
+module.exports = { scan, scanBuffer, scanStream, Verdict };